+++ src/components/quicklist.js
@@ -27,7 +27,7 @@ function addQuicklistPanelButtons() {
 }
 
 function updateSelectQuicklist() {
-    $("#bp-custom-select-ql").toggleClass("disabled", !unsafeWindow.selection_mode);
+    $("#bp-custom-select-ql").toggleClass("disabled", !Page.bp().selectionMode);
 }
 
 function onActionButtonClick() {
@@ -130,7 +130,7 @@ function addEventListeners() {
     $('.item:not(.spacer)').click(updateSelectQuicklist);
 
     $("#bp-custom-select-ql").click(function () {
-        if (unsafeWindow.selection_mode) selectQuicklist();
+        if (Page.bp().selectionMode) selectQuicklist();
     });
 }
 
@@ -142,7 +142,7 @@ function listSelection(value) {
         at = 0;
 
     _clearSelection();
-    unsafeWindow.updateClearSelectionState();
+    Page.bp().updateClearSelectionState();
 
     selection.each(function () {
         var $this = $(this);
@@ -251,28 +251,23 @@ function selectItem(element) { element.removeClass('unselected'); }
 function unselectItem(element) { element.addClass('unselected'); }
 
 function addSelectPage() {
+    var backpack = Page.bp();
+
     function selectItems(items) {
-        unsafeWindow.selection_mode = true;
+        backpack.selectionMode = true;
         selectItem(items);
 
-        unsafeWindow.updateClearSelectionState();
-        unsafeWindow.calculateValue();
+        backpack.updateClearSelectionState();
+        backpack.updateValues();
         updateSelectQuicklist();
     }
 
     $('#backpack').on('click', '.select-page', function () {
-        var page = +this.dataset.page,
-            pageitems;
-
-        if (page >= 1) {
-            pageitems = $('.pagenum[data-page-num="' + page + '"]').nextUntil('.pagenum').not('.spacer').filter(':visible');
-        } else { // new items
-            pageitems = $('#newlist .item');
-        }
+        var pageitems = $(this).closest('.backpack-page').find('.item').not('.spacer').filter(':visible');
 
         if (!pageitems.length) return;
 
-        if (unsafeWindow.selection_mode) {
+        if (backpack.selectionMode) {
             if (pageitems.length === pageitems.not('.unselected').length) { // all == selected
                 unselectItem(pageitems);
 
@@ -291,29 +286,25 @@ function addSelectPage() {
 }
 
 function _clearSelection() {
-    unsafeWindow.clearSelection();
+    Page.bp().clearSelection();
     updateSelectQuicklist();
 }
 
 function addSelectPageButtons() {
-    $('.pagenum').each(function () {
+    $('.page-number').each(function () {
         var $this = $(this),
             label = $this.find('.page-anchor'),
             page, sp;
 
         if (!label[0]) return;
-        page = label[0].id.replace('page', '');
         sp = $this.find('.select-page');
 
         if (sp.length) {
-            $this.attr('data-page-num', page);
-            sp.attr('date-page', page);
             return;
         }
 
-        if (!$this.nextUntil('.pagenum').not('.spacer').filter(':visible').length) return;
-        $this.attr('data-page-num', page);
-        label.after('<span class="btn btn-primary btn-xs pull-right select-page" data-page="' + page + '" style="margin-right: 2.7%;margin-top: -0.1%;">Select Page</span>');
+        if (!$this.nextUntil('.page-number').not('.spacer').filter(':visible').length) return;
+        label.after('<span class="btn btn-primary btn-xs pull-right select-page" style="margin-right: 2.7%;margin-top: -0.1%;">Select Page</span>');
     });
 }
 
@@ -324,9 +315,9 @@ function addHooks() {
         }
     });
 
-    Script.exec("var old_updateMargins = window.updateMargins;"+
+    Script.exec("var old_updateDisplay = window.backpack.updateDisplay;"+
                 addSelectPageButtons+
-                "window.updateMargins = function () { old_updateMargins(); addSelectPageButtons(); }");
+                "window.backpack.updateDisplay = function () { old_updateDisplay.call(this); addSelectPageButtons(); }");
 }
 
 function load() {