+++ suite.meta.js
@@ -3,9 +3,10 @@
 // @name         backpack.tf enhancement suite
 // @namespace    http://steamcommunity.com/id/caresx/
 // @author       cares
-// @version      1.3.4.1
+// @version      1.3.5
 // @description  Enhances your backpack.tf experience.
-// @match        *://*.backpack.tf/*
+// @include      /^https?://.*\.?backpack\.tf/.*$/
+// @exclude      /^https?://forums\.backpack\.tf/.*$/
 // @require      https://code.jquery.com/jquery-2.1.3.min.js
 // @require      https://maxcdn.bootstrapcdn.com/bootstrap/3.3.4/js/bootstrap.min.js
 // @require      https://cdn.rawgit.com/caresx/steam-econcc/6a2bc2d4abfab4fe9b46aae5ab07aa1db6a544cd/econcc.js